+++ packages/brick_offline_first_with_rest/lib/offline_queue.dart
@@ -1,0 +1,4 @@
+export 'package:brick_offline_first_with_rest/src/offline_queue/rest_offline_queue_client.dart';
+export 'package:brick_offline_first_with_rest/src/offline_queue/rest_offline_request_queue.dart';
+export 'package:brick_offline_first_with_rest/src/offline_queue/rest_request_sqlite_cache.dart';
+export 'package:brick_offline_first_with_rest/src/offline_queue/rest_request_sqlite_cache_manager.dart';