+++ packages/brick_offline_first_with_supabase_build/test/offline_first_generator_test.dart
@@ -60,6 +60,7 @@ Future<void> generateExpectation(
 }) async {
   final reader = await generateReader(filename);
   final generated = await (generator ?? _generator).generate(reader, MockBuildStep());
+  print(generated);
   expect(generated.trim(), output.trim());
 }