+++ packages/brick_offline_first_with_rest/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Expose offline queue functionality in `offline_queue.dart`
+
 ## 3.0.2
 
 * Apply standardized lints